Offline mode, advanced scrolling, easter egg

+++ ETGModder.cs
@@ -11,6 +11,8 @@ using System.Threading;
 namespace ETGModInstaller {
     public static class ETGModder {
 
+        public static bool IsOffline = false;
+
         public static string LogPath;
         public static string ExePath;
         public static string ExeBackupPath;
@@ -83,21 +85,25 @@ namespace ETGModInstaller {
             ins.Deobfuscate("Assembly-CSharp.dll");
 
             //Setup the files and MonoMod instances
-            ins.LogLine("Mod #0: API MOD (ETGMod)");
-            //Check if the revision online is newer
-            RepoHelper.RevisionFile = Path.Combine(ins.MainMod.Dir.FullName, "ModCache", "ETGMOD_REVISION.txt");
-            int revisionOnline = RepoHelper.RevisionOnline;
-            if (RepoHelper.Revision < revisionOnline) {
-                string cachedPath = Path.Combine(ins.MainMod.Dir.FullName, "ModCache", "ETGMOD.zip");
-                if (File.Exists(cachedPath)) {
-                    File.Delete(cachedPath);
+            int mi = -1;
+            if (!IsOffline) {
+                mi = 0;
+                ins.LogLine("Mod #0: Base");
+                //Check if the revision online is newer
+                RepoHelper.RevisionFile = Path.Combine(ins.MainMod.Dir.FullName, "ModCache", "ETGMOD_REVISION.txt");
+                int revisionOnline = RepoHelper.RevisionOnline;
+                if (RepoHelper.Revision < revisionOnline) {
+                    string cachedPath = Path.Combine(ins.MainMod.Dir.FullName, "ModCache", "ETGMOD.zip");
+                    if (File.Exists(cachedPath)) {
+                        File.Delete(cachedPath);
+                    }
                 }
+                if (!IsOffline && !ins.UnzipMod(ins.DownloadCached(RepoHelper.ETGModURL, "ETGMOD.zip"))) {
+                    return;
+                }
+                RepoHelper.Revision = revisionOnline;
             }
-            if (!ins.UnzipMod(ins.DownloadCached(RepoHelper.ETGModURL, "ETGMOD.zip"))) {
-                return;
-            }
-            RepoHelper.Revision = revisionOnline;
-            int mi = 0;
+            
 
             int[] selectedIndices = null;
             ins.Invoke(delegate() {
@@ -106,11 +112,14 @@ namespace ETGModInstaller {
                 selectedIndices = _selectedIndices;
             });
             while (selectedIndices == null) {
-                System.Threading.Thread.Sleep(100);
+                Thread.Sleep(100);
             }
 
             for (int i = 0; i < selectedIndices.Length; i++) {
                 Tuple<string, string> t = ins.APIMods[selectedIndices[i]];
+                if (string.IsNullOrEmpty(t.Item2)) {
+                    continue;
+                }
                 ins.Log("Mod #").Log((++mi).ToString()).Log(": ").LogLine(t.Item1);
                 if (!ins.UnzipMod(ins.DownloadCached(t.Item2, t.Item1 + ".zip"))) {
                     return;
@@ -311,8 +320,12 @@ namespace ETGModInstaller {
         }
         
         public static byte[] Download(this InstallerWindow ins, string url) {
+            if (IsOffline) {
+                return null;
+            }
+
             byte[] data = null;
-            
+
             ins.Log("Downloading ").Log(url).LogLine("...");
             ins.InitProgress("Starting download", 1);
             
@@ -388,6 +401,9 @@ namespace ETGModInstaller {
             }
             
             data = ins.Download(url);
+            if (data == null) {
+                return null;
+            }
             
             ins.WriteDataToCache(cached, data);
             return data;
@@ -461,6 +477,9 @@ namespace ETGModInstaller {
 		}
         
         public static bool UnzipMod(this InstallerWindow ins, byte[] data) {
+            if (data == null) {
+                return false;
+            }
             using (MemoryStream ms = new MemoryStream(data, 0, data.Length, false, true)) {
                 return ins.UnzipMod(ms);
             }
@@ -601,7 +620,8 @@ namespace ETGModInstaller {
             monomod.Out = monomod.In;
             using (FileStream fileStream = File.Open(LogPath, FileMode.Append)) {
                 using (StreamWriter streamWriter = new StreamWriter(fileStream)) {
-                    monomod.Logger = (string s) => streamWriter.WriteLine(s);
+                    monomod.Logger = (string s) => ins.OnActivity();
+                    monomod.Logger += (string s) => streamWriter.WriteLine(s);
                     try {
                         monomod.AutoPatch(true, true);
                         return true;
@@ -622,7 +642,8 @@ namespace ETGModInstaller {
 
             using (FileStream fileStream = File.Open(LogPath, FileMode.Append)) {
                 using (StreamWriter streamWriter = new StreamWriter(fileStream)) {
-                    ins.MainMod.Logger = (string s) => streamWriter.WriteLine(s);
+                    ins.MainMod.Logger = (string s) => ins.OnActivity();
+                    ins.MainMod.Logger += (string s) => streamWriter.WriteLine(s);
                     try {
                         ins.MainMod.AutoPatch(true, true);
                         return true;
